# docs/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Curator v2 API
  version: 1.10.0
  description: |
    C4 of the frozen contracts (design doc 03). JSON in, JSON out, UTF-8.
    Every route lives under the mount prefix `{base}` (empty by default, `/curation` in
    production; the gateway does not strip it). Write operations are asynchronous: they
    return the resource and its state, progress arrives over SSE or polling.
    Errors use one body, `Error`; `code` is stable for machines, `message` is Chinese for people.
    `precheck_failed` (422, any write that starts work) carries `error.details` shaped as
    `PrecheckDetails`: one entry per pre-start check (D30).
    Responses that point people at a page carry `links` (D22): absolute URLs when
    `server.publicBaseUrl` is configured, otherwise relative paths with `absolute: false`.
    Write endpoints accept only `application/json` - send the header on every write, even without
    a body - and cross-site writes are refused; there is no CORS.
    Ids are opaque strings. Records made since 1.10.0 get `<prefix>-<9 lowercase letters>` (task,
    sub, ds, pf, cred, vb, vm, upl; e.g. `task-kqzmrtbwe`, D45); records made before keep their
    `<prefix>_...` id, and both work everywhere. Ids carry no time: lists sort by creation time.

    1.1.0 (2026-09-21): dataset registry (D36): `/datasets` now lists registered datasets and the
    TOS / cache-bucket listing moved to `/datasets/browse`; `/overview`; starting a task re-checks
...

...

...

...

...

...

...

...

...

...

...
servers:
...
security:
  - basicAuth: []
...
paths:
...
  /datasets:
    get:
      tags: [datasets]
      operationId: listDatasets
      summary: Registered datasets, page-number pagination, newest first (D36)
      parameters:
        - name: page
          in: query
          schema: {type: integer, minimum: 1, default: 1}
        - name: page_size
          in: query
          schema: {enum: [10, 20, 50, 100], default: 20}
        - name: q
          in: query
          description: search by name or address
          schema: {type: string}
        - name: format
          in: query
          schema: {$ref: "#/components/schemas/DatasetFormat"}
        - name: check_state
          in: query
          schema: {enum: [ok, changed]}
      responses:
        "200":
          description: one page
          content:
            application/json:
              schema:
                type: object
                required: [items, page, page_size, total]
                properties:
                  items: {type: array, items: {$ref: "#/components/schemas/DatasetItem"}}
                  page: {type: integer, minimum: 1}
                  page_size: {type: integer}
                  total: {type: integer, minimum: 0}
        default: {$ref: "#/components/responses/Error"}
    post:
      tags: [datasets]
      operationId: createDataset
      summary: Register a dataset - preflight plus the full file listing, both fingerprints kept (D36)
      description: |
        Registering the same source + address + region again returns the existing registration
        with 200 instead of creating a second one.
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/DatasetCreate"}
      responses:
        "201":
          description: registered
          content:
            application/json:
              schema: {$ref: "#/components/schemas/DatasetDetail"}
        "200":
          description: already registered
          content:
            application/json:
              schema: {$ref: "#/components/schemas/DatasetDetail"}
        default: {$ref: "#/components/responses/Error"}
...
  /datasets/{id}:
    parameters:
      - $ref: "#/components/parameters/PathId"
    get:
      tags: [datasets]
      operationId: getDataset
      summary: One registration with its preflight result, recent fingerprint checks and tasks
      responses:
        "200":
          description: dataset
          content:
            application/json:
              schema: {$ref: "#/components/schemas/DatasetDetail"}
        default: {$ref: "#/components/responses/Error"}
    patch:
      tags: [datasets]
      operationId: updateDataset
      summary: Rename, or change the note
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/DatasetPatch"}
      responses:
        "200":
          description: updated
          content:
            application/json:
              schema: {$ref: "#/components/schemas/DatasetDetail"}
        default: {$ref: "#/components/responses/Error"}
    delete:
      tags: [datasets]
      operationId: deleteDataset
      summary: Remove the registration; the data on TOS is untouched. In use by an unfinished task -> 409 dataset_in_use
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "204": {description: removed}
        default: {$ref: "#/components/responses/Error"}
...
  /preflight:
    post:
      tags: [datasets]
      operationId: preflight
      summary: Preflight a dataset (metadata only, seconds); the result is kept 30 minutes
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/PreflightRequest"}
      responses:
        "200":
          description: preflight result
          content:
            application/json:
              schema: {$ref: "#/components/schemas/PreflightResponse"}
        default: {$ref: "#/components/responses/Error"}
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
      description: HTTP Basic against htpasswd (bcrypt / apr1), realm "Robot Data Curation" (doc 08 §5)
  parameters:
    PathId:
      name: id
      in: path
      required: true
      schema: {type: string}
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      description: the same key within 24 hours returns the first response (doc 03 §8)
      schema: {type: string, minLength: 8, maxLength: 128}
...
  responses:
    Error:
      description: |
        400 validation_failed, 401 unauthorized, 404 not_found, 409 task_state_conflict /
        subtask_active / credential_in_use / backend_in_use / name_taken / preflight_expired /
        source_changed / result_changed / idempotency_conflict, 412 precondition_failed,
        422 precheck_failed / confirm_path_mismatch / model_check_failed, 500 internal
      content:
        application/json:
          schema: {$ref: "#/components/schemas/Error"}
  schemas:
...
    Error:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message]
          properties:
            code:
              enum: [validation_failed, unauthorized, not_found, task_state_conflict, subtask_active,
                     credential_in_use, backend_in_use, dataset_in_use, name_taken, preflight_expired, precheck_failed,
                     source_changed, result_changed, confirm_path_mismatch, model_check_failed,
                     idempotency_conflict, precondition_failed, method_not_allowed, internal]
            message: {type: string, description: "Chinese, shown to people as is"}
            details: {type: object}
    Link:
      type: object
      additionalProperties: false
      required: [rel, title, url]
      properties:
        rel: {enum: [task, report, adjudication, plan, logs, credentials, dataset]}
        title: {type: string}
        url: {type: string}
        absolute: {type: boolean, default: true, description: "false when publicBaseUrl is not configured"}
    Links:
      type: array
      items: {$ref: "#/components/schemas/Link"}
...
    Region:
      type: string
      pattern: "^[a-z0-9][a-z0-9-]{1,31}$"
...
    DatasetFormat: {enum: [lerobot_v2, lerobot_v3, unsupported]}
    DatasetItemFields:
      type: object
      required: [id, name, source, uri, region, format, episode_count, robot_type, check_state, checked_at,
                 preflighted_at, created_at, last_task]
      properties:
        id:
          type: string
          pattern: "^(ds-[a-z]{9}|ds_[0-9A-Za-z]+)$"
          description: "ds- and 9 lowercase letters (D45); registrations made before 1.10.0 keep ds_ and their old id"
        name: {type: string}
        source: {enum: [tos, public, local]}
        uri: {type: string}
        region: {type: [string, "null"]}
        format: {$ref: "#/components/schemas/DatasetFormat"}
        episode_count: {type: [integer, "null"], minimum: 0}
        robot_type: {type: [string, "null"]}
        check_state: {enum: [ok, changed], description: "changed: the fingerprints differ from the kept ones; preflight again"}
        checked_at: {type: [integer, "null"]}
        preflighted_at: {type: integer}
        created_at: {type: integer}
        last_task:
          oneOf:
            - type: "null"
            - $ref: "#/components/schemas/TaskRef"
    DatasetItem:
      allOf:
        - $ref: "#/components/schemas/DatasetItemFields"
      unevaluatedProperties: false
    DatasetDetail:
      allOf:
        - $ref: "#/components/schemas/DatasetItemFields"
        - type: object
          required: [note, credential, preflight, meta_fingerprint, listing, checks, tasks, links]
          properties:
            note: {type: [string, "null"]}
            credential: {type: [string, "null"], description: "access key name; null for the public bucket, or once the key is deleted"}
            preflight: {$ref: "./cli/preflight.schema.json"}
            meta_fingerprint: {type: string}
            listing:
              type: object
              description: summary of the file listing kept at the last preflight (source-manifest summary)
              required: [objects, bytes, digest]
              properties:
                objects: {type: integer, minimum: 0, description: "C2 source-manifest summary.count"}
                bytes: {type: integer, minimum: 0}
                digest: {type: string}
            checks: {type: array, maxItems: 20, items: {$ref: "#/components/schemas/DatasetCheck"}, description: "newest first"}
            tasks: {type: array, maxItems: 20, items: {$ref: "#/components/schemas/TaskRef"}, description: "newest first"}
            links: {$ref: "#/components/schemas/Links"}
      unevaluatedProperties: false
    DatasetCreate:
      type: object
      additionalProperties: false
      required: [input]
      properties:
        input: {$ref: "#/components/schemas/InputRef"}
        name: {type: string, minLength: 1, maxLength: 128, description: "defaults to the last path segment"}
        note: {type: string, maxLength: 2000}
    DatasetPatch:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        name: {type: string, minLength: 1, maxLength: 128}
        note: {type: [string, "null"], maxLength: 2000}
    DatasetCheck:
      type: object
      additionalProperties: false
      required: [at, trigger, result, change]
      properties:
        at: {type: integer}
        trigger: {enum: [add, recheck, task_start, repreflight]}
        result: {enum: [same, changed]}
        change:
          oneOf:
            - type: "null"
            - $ref: "#/components/schemas/SourceChange"
    SourceChange:
      type: object
      additionalProperties: false
      description: what differs from the fingerprints kept at registration or at the last preflight (D37)
      required: [meta_changed, added, removed, modified, sample_keys, preflighted_at]
      properties:
        meta_changed: {type: boolean}
        added: {type: integer, minimum: 0}
        removed: {type: integer, minimum: 0}
        modified: {type: integer, minimum: 0}
        sample_keys: {type: array, maxItems: 20, items: {type: string}, description: "a few affected keys, added ones first"}
        preflighted_at: {type: integer}
    TaskRef:
      type: object
      additionalProperties: false
      required: [id, name, state, created_at]
      properties:
        id: {type: string}
        name: {type: string}
        state: {$ref: "#/components/schemas/TaskState"}
        created_at: {type: integer}
...
    InputRef:
      type: object
      additionalProperties: false
      required: [source, uri]
      properties:
        source: {enum: [tos, public, local]}
        uri: {type: string, minLength: 1}
        region: {$ref: "#/components/schemas/Region"}
        credential: {type: [string, "null"], description: "access key name; not for source=public. Requests name one; responses give null once the key was deleted (see rebind-credentials)"}
      if:
        properties: {source: {const: tos}}
      then:
        required: [credential]
    InputSpec:
      description: a registered dataset (D36), or an input given in full
      oneOf:
        - type: object
          additionalProperties: false
          required: [dataset_id]
          properties:
            dataset_id: {type: string}
        - $ref: "#/components/schemas/InputRef"
...
    PreflightRequest:
      type: object
      additionalProperties: false
      required: [input]
      properties:
        input: {$ref: "#/components/schemas/InputSpec"}
        vlm_backend: {type: string}
        embodiment_id: {type: string}
    PreflightResponse:
      type: object
      additionalProperties: false
      required: [preflight_id, expires_at, result]
      properties:
        preflight_id: {type: string}
        expires_at: {type: integer}
        result: {$ref: "./cli/preflight.schema.json"}
...
    TaskState:
      enum: [created, queued, running, pausing, paused, stopping, stopped, succeeded, completed_with_errors, failed]
...